+++ khiops/samples/samples_sklearn.py
@@ -57,14 +57,6 @@ def khiops_classifier():
     # Train the classifier
     khc.fit(X_train, y_train)
 
-    # Show the feature importance info
-    print(f"Features evaluated: {khc.n_features_evaluated_}")
-    print(f"Features selected : {khc.n_features_used_}")
-    print("Top 3 used features")
-    for i, feature in enumerate(khc.feature_used_names_[:3]):
-        print(f"{feature} - Importance: {khc.feature_used_importances_[i][2]}")
-    print("---")
-
     # Predict the classes on the test dataset
     y_test_pred = khc.predict(X_test)
     print("Predicted classes (first 10):")
@@ -180,14 +172,6 @@ def khiops_classifier_text():
 
     # Train the classifier
     khc.fit(X_train, y_train)
-
-    # Show the feature importance info
-    print(f"Features evaluated: {khc.n_features_evaluated_}")
-    print(f"Features selected : {khc.n_features_used_}")
-    print("Top 3 used features")
-    for i, feature in enumerate(khc.feature_used_names_[:3]):
-        print(f"{feature} - Importance: {khc.feature_used_importances_[i][2]}")
-    print("---")
 
     # Predict the classes on the test dataset
     y_test_pred = khc.predict(X_test)
@@ -309,14 +293,6 @@ def khiops_classifier_multitable_snowflake():
     # Train the classifier (by default it creates 1000 multi-table features)
     khc = KhiopsClassifier(n_trees=0)
     khc.fit(X_train, y_train)
-
-    # Show the feature importance info
-    print(f"Features evaluated: {khc.n_features_evaluated_}")
-    print(f"Features selected : {khc.n_features_used_}")
-    print("Top 3 used features")
-    for i, feature in enumerate(khc.feature_used_names_[:3]):
-        print(f"{feature} - Importance: {khc.feature_used_importances_[i][2]}")
-    print("---")
 
     # Predict the class on the test dataset
     y_test_pred = khc.predict(X_test)
@@ -550,14 +526,6 @@ def khiops_regressor():
     # Train the regressor
     khr.fit(X_train, y_train)
 
-    # Show the feature importance info
-    print(f"Features evaluated: {khr.n_features_evaluated_}")
-    print(f"Features selected : {khr.n_features_used_}")
-    print("Top 3 used features")
-    for i, feature in enumerate(khr.feature_used_names_[:3]):
-        print(f"{feature} - Importance: {khr.feature_used_importances_[i][2]}")
-    print("---")
-
     # Predict the values on the test dataset
     y_test_pred = khr.predict(X_test)
     print("Predicted values for 'age' (first 10):")
@@ -695,13 +663,6 @@ def khiops_encoder_multitable_snowflake():
     # Create the KhiopsEncoder with 10 additional multitable features and fit it
     khe = KhiopsEncoder(n_features=10)
     khe.fit(X, y)
-
-    # Show the feature importance info
-    print(f"Features evaluated: {khe.n_features_evaluated_}")
-    print("Top 3 evaluated features")
-    for i, feature in enumerate(khe.feature_evaluated_names_[:3]):
-        print(f"{feature} - Level: {khe.feature_evaluated_importances_[i]}")
-    print("---")
 
     # Transform the train dataset
     print("Encoded feature names:")